+++ CLAM/src/Visualization/QTVM/Plots/SingleDisplayPlot.hxx
@@ -70,7 +70,7 @@ namespace CLAM
 			// signals from SegmentEditor
 			void segmentOnsetChanged(unsigned,double);
 			void segmentOffsetChanged(unsigned,double);
-			void segmentInserted(double);
+			void segmentInserted(unsigned);
 			void segmentDeleted(unsigned);
 			void currentSegmentChanged(unsigned);
